Split the example code
